requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
numpy==1.24.3
gunicorn==21.2.0
